+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
   * gpxdo split --waypoints
+  * gpxdo split --points